backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncio==3.4.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
nba_api==1.4.1
numpy==1.26.1
pandas==2.1.2
pipreqs==0.4.13
pydantic==2.4.2
pydantic_core==2.10.1
pyhumps==3.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarg==0.1.9
